:hammer: Refactor domain classes directory
