+++ app/src/main/java/com/rummenigged/popcorntime/domain/model/Series.kt
@@ -1,4 +1,4 @@
-package com.rummenigged.popcorntime.domain
+package com.rummenigged.popcorntime.domain.model
 
 data class Series(
     val id: Int,
