+++ app/src/main/java/com/rummenigged/popcorntime/domain/model/SeriesSearchResult.kt
@@ -1,4 +1,4 @@
-package com.rummenigged.popcorntime.domain
+package com.rummenigged.popcorntime.domain.model
 
 data class SeriesSearchResult(
     val id: Int,
